backend/requirements.txt
fastapi
uvicorn
torch
pillow
opencv-python
git+https://github.com/openai/CLIP.git
numpy
scikit-learn
